# contracts/openapi/agentintel.openapi.yaml
openapi: 3.1.0
info:
  title: AGENTintel Local API
  version: 1.0.0
  description: >-
    Evidence-first API for local competitive research. Mutating operations are
    asynchronous and return a durable run identifier.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:7465
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /v1/session/bootstrap:
    post:
      operationId: sessionBootstrap
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionBootstrapRequest"
      responses:
        "200":
          description: One-time dashboard ticket exchanged for an HttpOnly session
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/session:
    get:
      operationId: sessionGet
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current dashboard session and CSRF token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/datasets/competitive-pulse/preview:
    post:
      operationId: datasetPreview
      requestBody:
        required: true
        content:
          text/csv; charset=utf-8:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: Completed semantic preview; valid input becomes a durable opaque dataset.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "400": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
  /v1/datasets/{datasetId}:
    parameters:
      - $ref: "#/components/parameters/DatasetId"
    get:
      operationId: datasetGet
      responses:
        "200":
          description: Content-free imported dataset summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: datasetDelete
      responses:
        "202":
          description: Dataset deletion gate accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "404": { $ref: "#/components/responses/Problem" }
  /v1/comparisons:
    post:
      operationId: comparisonStart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComparisonStartRequest"
      responses:
        "202":
          description: Comparison accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/research:
    post:
      operationId: researchStart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchStartRequest"
      responses:
        "202":
          description: Research accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/runs:
    get:
      operationId: runList
      parameters:
        - in: query
          name: project_id
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Recent runs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Run" }
  /v1/runs/{runId}:
    get:
      operationId: runGet
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Durable run state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunDetail" }
        "404": { $ref: "#/components/responses/Problem" }
  /v1/runs/{runId}/events:
    get:
      operationId: runEventsStream
      parameters:
        - $ref: "#/components/parameters/RunId"
        - in: header
          name: Last-Event-ID
          schema: { type: string }
      responses:
        "200":
          description: Ordered run event stream
          content:
            text/event-stream:
              schema: { type: string }
  /v1/runs/{runId}/cancel:
    post:
      operationId: runCancel
      parameters:
        - $ref: "#/components/parameters/RunId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason: { type: string, maxLength: 500 }
      responses:
        "202":
          description: Cancellation requested
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "409": { $ref: "#/components/responses/Problem" }
  /v1/runs/{runId}/replay:
    post:
      operationId: runReplay
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: Replay accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "409": { $ref: "#/components/responses/Problem" }
  /v1/runs/{runId}/report:
    get:
      operationId: reportGet
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Cited comparison report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ComparisonReport" }
        "404": { $ref: "#/components/responses/Problem" }
  /v1/search:
    get:
      operationId: searchGet
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 2, maxLength: 200 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Evidence-backed local search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SearchResult" }
  /v1/entities/{entityId}:
    get:
      operationId: entityGet
      parameters:
        - in: path
          name: entityId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Entity header and observed identifiers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-token
    sessionCookie:
      type: apiKey
      in: cookie
      name: agentintel_session
  parameters:
    RunId:
      in: path
      name: runId
      required: true
      schema: { type: string, minLength: 1 }
    DatasetId:
      in: path
      name: datasetId
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    SessionBootstrapRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 43
          maxLength: 43
          pattern: "^[A-Za-z0-9_-]{43}$"
    Session:
      type: object
      additionalProperties: false
      required: [csrf, expires_at]
      properties:
        csrf: { type: string, minLength: 32 }
        expires_at: { type: string, format: date-time }
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
    ComparisonStartRequest:
      type: object
      additionalProperties: false
      required: [project_id, target_ids]
      properties:
        project_id: { type: string, minLength: 1, maxLength: 100 }
        target_ids:
          type: array
          minItems: 2
          maxItems: 50
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 100 }
        connector_ids:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string }
          default: [fixture.competitive-pulse]
        dataset_id: { type: string, minLength: 1, maxLength: 200 }
        goal: { type: string, maxLength: 1000 }
        simulate:
          enum: [none, source_failure, corrupt_artifact, slow]
          default: none
    ResearchStartRequest:
      type: object
      additionalProperties: false
      required: [project_id, question, target_ids]
      properties:
        project_id: { type: string, minLength: 1, maxLength: 100 }
        question: { type: string, minLength: 3, maxLength: 2000 }
        target_ids:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 100 }
        source_budget: { type: integer, minimum: 1, maximum: 100, default: 20 }
    ImportInputSummary:
      type: object
      additionalProperties: false
      required: [schema_id, sha256, size_bytes, row_count]
      properties:
        schema_id: { const: "agentintel.competitive-pulse-import.v1" }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        size_bytes: { type: integer, minimum: 0, maximum: 8388608 }
        row_count: { type: [integer, "null"], minimum: 0 }
    ImportPolicySummary:
      type: object
      additionalProperties: false
      required:
        [
          attestation_version,
          target_scope,
          data_class,
          permitted_purpose,
          retention_days,
          rights_state,
        ]
      properties:
        attestation_version:
          { const: "public-permitted-brand-competitive-research.v1" }
        target_scope: { const: brand }
        data_class: { const: public }
        permitted_purpose: { const: competitive_research }
        retention_days: { type: integer, minimum: 1, maximum: 365 }
        rights_state: { const: permitted }
    ImportTargetSummary:
      type: object
      additionalProperties: false
      required: [target_id, target_name, row_count, metric_availability]
      properties:
        target_id: { type: string }
        target_name: { type: string }
        row_count: { type: integer, minimum: 1 }
        metric_availability:
          type: object
          additionalProperties: false
          required:
            [
              followers.delta,
              public-engagement-by-followers.median,
              posting-cadence,
              content-format-mix,
            ]
          properties:
            followers.delta:
              { enum: [missing, insufficient, contradictory, available] }
            public-engagement-by-followers.median:
              { enum: [missing, insufficient, contradictory, available] }
            posting-cadence:
              { enum: [missing, insufficient, contradictory, available] }
            content-format-mix:
              { enum: [missing, insufficient, contradictory, available] }
    ImportDiagnostic:
      type: object
      additionalProperties: false
      required: [severity, code, record_number, column, message]
      properties:
        severity: { enum: [error, warning] }
        code: { type: string }
        record_number: { type: [integer, "null"], minimum: 1 }
        column: { type: [string, "null"] }
        message: { type: string }
    ImportPreview:
      type: object
      additionalProperties: false
      required:
        [
          schema_version,
          valid,
          input,
          targets,
          diagnostics,
          diagnostics_truncated,
        ]
      properties:
        schema_version: { const: "agentintel.import-preview.v1" }
        valid: { type: boolean }
        dataset_id: { type: string }
        state: { enum: [ready, deleting, deleted] }
        validated_at: { type: string, format: date-time }
        retention_until: { type: string, format: date-time }
        input: { $ref: "#/components/schemas/ImportInputSummary" }
        policy: { $ref: "#/components/schemas/ImportPolicySummary" }
        platform: { type: string }
        targets:
          type: array
          items: { $ref: "#/components/schemas/ImportTargetSummary" }
        diagnostics:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/ImportDiagnostic" }
        diagnostics_truncated: { type: boolean }
...
    Run:
      description: >-
        Durable run header. Provenance and immutable-input fields appear after
        the corresponding snapshot or successful derivation has been recorded.
      allOf:
        - $ref: "#/components/schemas/RunFields"
      unevaluatedProperties: false
    RunDetail:
      description: Durable run header with its ordered events and committed artifacts.
      allOf:
        - $ref: "#/components/schemas/RunFields"
        - type: object
          required: [events, artifacts]
          properties:
            events:
              type: array
              items: { $ref: "#/components/schemas/RunEvent" }
            artifacts:
              type: array
              items: { $ref: "#/components/schemas/Artifact" }
      unevaluatedProperties: false
    RunEvent:
      type: object
      additionalProperties: false
      required:
        [id, run_id, sequence, stage, level, message, progress, recorded_at]
      properties:
        id: { type: string }
        run_id: { type: string }
        sequence: { type: integer, minimum: 1 }
        stage: { type: string }
        level: { enum: [debug, info, warning, error] }
        message: { type: string }
        progress: { type: number, minimum: 0, maximum: 1 }
        recorded_at: { type: string, format: date-time }
...
    ComparisonReport:
      oneOf:
        - $ref: "#/components/schemas/ComparisonReportV1"
        - $ref: "#/components/schemas/ImportComparisonReport"
...
    SearchResult:
      type: object
      additionalProperties: false
      required: [kind, id, label, excerpt, confidence]
      properties:
        kind: { enum: [entity, report, claim, observation] }
        id: { type: string }
        label: { type: string }
        excerpt: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        run_id: { type: [string, "null"] }
    Entity:
      type: object
      additionalProperties: false
      required:
        [id, type, display_name, resolution_state, identifiers, updated_at]
      properties:
        id: { type: string }
        type:
          {
            enum:
              [
                organization,
                brand,
                creator,
                professional_profile,
                social_account,
              ],
          }
        display_name: { type: string }
        resolution_state: { enum: [observed, candidate, confirmed, rejected] }
        identifiers:
          type: array
          items:
            type: object
            required: [scheme, value, source]
            properties:
              scheme: { type: string }
              value: { type: string }
              source: { type: string }
        updated_at: { type: string, format: date-time }
...